LayoutTests/fast/scrolling/ios/autoscroll-input-when-very-zoomed-expected.txt

This test focuses a form, them zooms and scrolls the page. Then text is entered in the form, and we check to make sure the page has scrolled so that the input is visible again, but not more that once, as it should be visible after the first scroll.
PASS: page has scrolled when the selection is not visible, and not scrolled when the selection is already visible.
